FÁZIS 2 SOK kért módosítás

+++ admin/inventory.php
@@ -5,7 +5,7 @@ $selector ="inventory";
 require("../common/header.php");
 
 sqlExecute(
-		palletSQL(),
+		inventorySQL(),
 		'inventoryTable');
 
 require("../common/footer.php");
@@ -16,11 +16,11 @@ function inventoryTable($results){
 	print '<tr>';
 	print '<th>ID</th>';
 	print '<th>';
-	
 	print 'Alapanyag név</th>';
 	print '<th>Beszállító Neve</th>';
 	print '<th>Beérkezés ideje</th>';
 	print '<th>Mennyiség</th>';
+	print '<th>Bevételezési<br/>mennyiség</th>';
 	print '<th>Raktáros</th>';
 	print '<th>Módosít</th>';
 	print '<th>Töröl</th>';
@@ -28,17 +28,32 @@ function inventoryTable($results){
 	print '</thead>';
 	while($row = $results->fetch_assoc()) {
 			
-	
-		print '<tr>';
+		if($row["deleted"] == "0"){
+			if((int)$row["rest"] <= 0){
+				print '<tr class="zerorow">';
+			}else{
+				print '<tr>';
+			}
+		}else{
+			print '<tr class="deletedrow">';
+		}
 		print '<td>'.$row["id"].'</td>';
 		print '<td>'.$row["product"].'</td>';
 		print '<td>'.$row["supplier"].'</td>';
 		print '<td>'.$row["time"].'</td>';
 		print '<td>'.$row["rest"].'</td>';
+		print '<td>'.$row["origamount"].'</td>';
 		print '<td>'.$row["user"].'</td>';
-		print '<td><button class="btn btn-sm btn-default" onclick="inventory_update('.$row["id"].','.$row["rest"].",'".$row["time"]."'".','.$row["origamount"].')">Szerkeszt</button></td>';
-		print '<td><button class="btn btn-sm btn-danger" onclick="deletePallet('.$row["id"].')">Töröl</button></td>';
-		print '</tr>';
+		
+		if($row["deleted"] == "0"){
+			print '<td><button class="btn btn-sm btn-default" onclick="inventory_update('.$row["id"].','.$row["rest"].",'".$row["time"]."'".','.$row["origamount"].')">Szerkeszt</button></td>';
+			print '<td><button class="btn btn-sm btn-danger" onclick="deletePallet('.$row["id"].')">Töröl</button></td>';
+		}else{
+			print '<td><button class="btn btn-sm btn-default" onclick="inventory_update('.$row["id"].','.$row["rest"].",'".$row["time"]."'".','.$row["origamount"].')">Szerkeszt</button></td>';
+			print '<td><button class="btn btn-sm btn-danger" onclick="undeletePallet('.$row["id"].')">Visszaállít</button></td>';
+				
+		}
+				print '</tr>';
 	}
 	print '</table>';
 }
